requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
anthropic==0.39.0
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.13.1
